refactor: 유틸함수 명칭 변경
